Case 24549: Finished InitializeDepths Cuda conversion from CSharp code; Still refactoring is not done yet

+++ Strict.Compiler.Cuda.Tests/Input/AddNumbers.cs
@@ -1,6 +1,6 @@
 ﻿public class AddNumbers
 {
-	public int Add(int first, int second)
+	public float Add(float first, float second)
 	{
 		return first + second;
 	}
